Rename CategoryAutocomplete to be consistent with other autocomplete components
