+++ packages/desktop-client/src/components/autocomplete/CategoryAutocomplete.tsx
