+++ packages/desktop-client/src/components/util/GenericInput.js
@@ -5,7 +5,7 @@ import { getMonthYearFormat } from 'loot-core/src/shared/months';
 
 import AccountAutocomplete from '../autocomplete/AccountAutocomplete';
 import Autocomplete from '../autocomplete/Autocomplete';
-import CategoryAutocomplete from '../autocomplete/CategorySelect';
+import CategoryAutocomplete from '../autocomplete/CategoryAutocomplete';
 import PayeeAutocomplete from '../autocomplete/PayeeAutocomplete';
 import SavedFilterAutocomplete from '../autocomplete/SavedFilterAutocomplete';
 import Input from '../common/Input';
